# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PhoneNumberAnswerMode = typeof PhoneNumberAnswerMode[keyof typeof PhoneNumberAnswerMode];


export const PhoneNumberAnswerMode = {
  forward: 'forward',
  ai_voice: 'ai_voice',
  voicemail: 'voicemail',
  reject: 'reject',
} as const;

export interface PhoneNumber {
  id: number;
  number: string;
  /** @nullable */
  twilioSid?: string | null;
  friendlyName: string;
  callerIdName: string;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  forwardTo: string | null;
  ringCount: number;
  answerMode: PhoneNumberAnswerMode;
  /** @nullable */
  aiSystemPrompt?: string | null;
  /** @nullable */
  voicemailGreeting?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type PhoneNumberInputAnswerMode = typeof PhoneNumberInputAnswerMode[keyof typeof PhoneNumberInputAnswerMode];


export const PhoneNumberInputAnswerMode = {
  forward: 'forward',
  ai_voice: 'ai_voice',
  voicemail: 'voicemail',
  reject: 'reject',
} as const;

export interface PhoneNumberInput {
  number: string;
  /** @nullable */
  twilioSid?: string | null;
  friendlyName: string;
  callerIdName: string;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  forwardTo?: string | null;
  ringCount?: number;
  answerMode: PhoneNumberInputAnswerMode;
  /** @nullable */
  aiSystemPrompt?: string | null;
  /** @nullable */
  voicemailGreeting?: string | null;
}

/**
 * @nullable
 */
export type PhoneNumberUpdateAnswerMode = typeof PhoneNumberUpdateAnswerMode[keyof typeof PhoneNumberUpdateAnswerMode] | null;


export const PhoneNumberUpdateAnswerMode = {
  forward: 'forward',
  ai_voice: 'ai_voice',
  voicemail: 'voicemail',
  reject: 'reject',
} as const;

export interface PhoneNumberUpdate {
  /** @nullable */
  friendlyName?: string | null;
  /** @nullable */
  callerIdName?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  forwardTo?: string | null;
  /** @nullable */
  ringCount?: number | null;
  /** @nullable */
  answerMode?: PhoneNumberUpdateAnswerMode;
  /** @nullable */
  aiSystemPrompt?: string | null;
  /** @nullable */
  voicemailGreeting?: string | null;
  /** @nullable */
  isActive?: boolean | null;
}

export type AvailableNumberCapabilities = {
  voice?: boolean;
  sms?: boolean;
  mms?: boolean;
};

export interface AvailableNumber {
  phoneNumber: string;
  friendlyName: string;
  /** @nullable */
  locality: string | null;
  /** @nullable */
  region: string | null;
  /** @nullable */
  postalCode?: string | null;
  isoCountry: string;
  capabilities?: AvailableNumberCapabilities;
}

export interface TestCallInput {
  toNumber: string;
}

export interface TestCallResult {
  callSid: string;
  status: string;
}

export interface Contact {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface ContactInput {
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
}

export interface ContactUpdate {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
}

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CompanyInput {
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface CompanyUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type CallLogDirection = typeof CallLogDirection[keyof typeof CallLogDirection];


export const CallLogDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export interface CallLog {
  id: number;
  /** @nullable */
  phoneNumberId?: number | null;
  /** @nullable */
  twilioCallSid?: string | null;
  direction: CallLogDirection;
  status: string;
  fromNumber: string;
  toNumber: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  recordingUrl?: string | null;
  /** @nullable */
  recordingSid?: string | null;
  /** @nullable */
  transcription?: string | null;
  /** @nullable */
  contactId?: number | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  callerIdName?: string | null;
  /** @nullable */
  answerMode?: string | null;
  createdAt: string;
}

export interface RecordingUrl {
  url: string;
  /** @nullable */
  expiresAt?: string | null;
}

export type AiVoiceConfigVoice = typeof AiVoiceConfigVoice[keyof typeof AiVoiceConfigVoice];


export const AiVoiceConfigVoice = {
  alloy: 'alloy',
  echo: 'echo',
  fable: 'fable',
  onyx: 'onyx',
  nova: 'nova',
  shimmer: 'shimmer',
} as const;

export interface AiVoiceConfig {
  id: number;
  voice: AiVoiceConfigVoice;
  greeting: string;
  systemPrompt: string;
  maxCallDuration?: number;
}

/**
 * @nullable
 */
export type AiVoiceConfigUpdateVoice = typeof AiVoiceConfigUpdateVoice[keyof typeof AiVoiceConfigUpdateVoice] | null;


export const AiVoiceConfigUpdateVoice = {
  alloy: 'alloy',
  echo: 'echo',
  fable: 'fable',
  onyx: 'onyx',
  nova: 'nova',
  shimmer: 'shimmer',
} as const;

export interface AiVoiceConfigUpdate {
  /** @nullable */
  voice?: AiVoiceConfigUpdateVoice;
  /** @nullable */
  greeting?: string | null;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  maxCallDuration?: number | null;
}

export interface DashboardStats {
  totalNumbers: number;
  activeNumbers: number;
  totalCalls: number;
  callsToday: number;
  avgDuration: number;
  totalContacts: number;
  totalCompanies: number;
  inboundCalls: number;
  outboundCalls: number;
  aiAnswered: number;
  voicemailCount: number;
  forwardedCalls: number;
}

export type SearchAvailableNumbersParams = {
/**
 * @nullable
 */
areaCode?: string | null;
/**
 * @nullable
 */
contains?: string | null;
country?: string;
/**
 * @nullable
 */
tollFree?: boolean | null;
};

export type ListContactsParams = {
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
companyId?: number | null;
};

export type ListCallLogsParams = {
/**
 * @nullable
 */
phoneNumberId?: number | null;
/**
 * @nullable
 */
direction?: ListCallLogsDirection;
/**
 * @nullable
 */
status?: string | null;
limit?: number;
};

export type ListCallLogsDirection = typeof ListCallLogsDirection[keyof typeof ListCallLogsDirection] | null;


export const ListCallLogsDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;

export type GetRecentCallsParams = {
limit?: number;
};