# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Call Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List CRM contacts
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /call-logs:
    get:
      operationId: listCallLogs
      tags: [call-logs]
      summary: List call logs with optional filters
      parameters:
        - name: phoneNumberId
          in: query
          schema:
            type: ["integer", "null"]
        - name: direction
          in: query
          schema:
            type: ["string", "null"]
            enum: [inbound, outbound, null]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallLog"
  /call-logs/{id}:
    get:
      operationId: getCallLog
      tags: [call-logs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
  /call-logs/{id}/recording:
    get:
      operationId: getRecordingUrl
      tags: [call-logs]
      summary: Get a signed recording URL for a call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recording URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingUrl"
...
components:
  schemas:
...
    Contact:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        industry:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CompanyUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CallLog:
      type: object
      required: [id, direction, status, fromNumber, toNumber, createdAt]
      properties:
        id:
          type: integer
        phoneNumberId:
          type: ["integer", "null"]
        twilioCallSid:
          type: ["string", "null"]
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
        fromNumber:
          type: string
        toNumber:
          type: string
        duration:
          type: ["integer", "null"]
        recordingUrl:
          type: ["string", "null"]
        recordingSid:
          type: ["string", "null"]
        transcription:
          type: ["string", "null"]
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        callerIdName:
          type: ["string", "null"]
        answerMode:
          type: ["string", "null"]
        createdAt:
          type: string
    RecordingUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string
        expiresAt:
          type: ["string", "null"]
...