# lib/api-zod/src/generated/types/callLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */
import type { CallLogDirection } from './callLogDirection';

export interface CallLog {
  id: number;
  /** @nullable */
  phoneNumberId?: number | null;
  /** @nullable */
  twilioCallSid?: string | null;
  direction: CallLogDirection;
  status: string;
  fromNumber: string;
  toNumber: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  recordingUrl?: string | null;
  /** @nullable */
  recordingSid?: string | null;
  /** @nullable */
  transcription?: string | null;
  /** @nullable */
  contactId?: number | null;
  /** @nullable */
  contactName?: string | null;
  /** @nullable */
  callerIdName?: string | null;
  /** @nullable */
  answerMode?: string | null;
  createdAt: string;
}