# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  firstName: string;
  lastName: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
}