# lib/api-zod/src/generated/types/recordingUrl.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Call Center API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordingUrl {
  url: string;
  /** @nullable */
  expiresAt?: string | null;
}